+++ ReleaseNotes/release_notes.md
@@ -21,7 +21,8 @@
   1. add meaningful `description` argument to all `progress_bar` instance
   1. extend `gof1d` and `gofnd` tests 
   1. Add `RooAddPdf::fixCoefNormalization( vars )` for all appearences of `RooAddPdf`
-    
+  1. Add (fictive) contetx mnager protocol to WorkManager
+  
 ## Backward incompatible
 
 ## Bug fixes